+++ Smithtry1000/circuitWidget.h
@@ -16,10 +16,10 @@ public:
     void addSvg(QString, int, int);
     void removeLastSvg();
 private:
-    void rImpedanceRealCalculation(float x, float y);
-    void rAdmitanceRealCalculation(float x, float y);
-    void rImpedanceImagCalculation(float x, float y);
-    void rAdmitanceImagCalculation(float x, float y);
+    void rImpedanceRealCalculation(double x, double y);
+    void rAdmitanceRealCalculation(double x, double y);
+    void rImpedanceImagCalculation(double x, double y);
+    void rAdmitanceImagCalculation(double x, double y);
     QList<int> tuned;
     QList<QString> paths;
     CircuitElements* circuitElements;
